Refactor HttpFuture to be a class, and re-introduce the callbacks
