+++ bravado/fido_client.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
 #
@@ -7,28 +6,50 @@
 
 import logging
 
+import fido
 from yelp_uri import urllib_utf8
 
-from bravado.fido_future import FidoFuture
 from bravado.mapping.http_client import APP_FORM, HttpClient
-from bravado.multipart_response import create_multipart_content
+from bravado.mapping.http_future import HttpFuture
 from bravado.mapping.param import stringify_body as param_stringify_body
+from bravado.mapping.response import ResponseLike
+from bravado.multipart_response import create_multipart_content
 
 log = logging.getLogger(__name__)
+
+
+class FidoResponseAdapter(ResponseLike):
+    """Wraps a fido.fido.Response object to provider a uniform interface
+    to the response innards.
+
+    :type requests_lib_response: :class:`fido.fido.Response`
+    """
+    def __init__(self, requests_lib_response):
+        self._delegate = requests_lib_response
+
+    @property
+    def status_code(self):
+        return self._delegate.code
+
+    def json(self, **_):
+        return self._delegate.json()
 
 
 class FidoClient(HttpClient):
     """Fido HTTP client implementation.
     """
 
-    def request(self, request_params, op=None):
+    def request(self, request_params, response_callback=None):
         """Sets up the request params as per Twisted Agent needs.
         Sets up crochet and triggers the API request in background
 
         :param request_params: request parameters for API call
         :type request_params: dict
+        :param response_callback: Function to be called after
+        receiving the response
+        :type response_callback: method
 
-        :return: crochet EventualResult
+        :return: :class: `bravado.mapping.http_future.HttpFuture`
         """
         url = '%s?%s' % (request_params['url'], urllib_utf8.urlencode(
             request_params.get('params', []), True))
@@ -39,9 +60,9 @@ class FidoClient(HttpClient):
             'headers': request_params.get('headers', {}),
         }
 
-        fido_future = FidoFuture(op, url, request_params)
-        fido_future.fetch()
-        return fido_future
+        return HttpFuture(fido.fetch(url, **request_params),
+                          FidoResponseAdapter,
+                          response_callback)
 
 
 def stringify_body(request_params):
