+++ bravado/requests_client.py
@@ -1,22 +1,71 @@
-#!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
 #
-# Copyright (c) 2013, Digium, Inc.
-# Copyright (c) 2014, Yelp, Inc.
+# Copyright (c) 2015, Yelp, Inc.
 #
 
 import logging
+import urlparse
 
 import requests
 import requests.auth
 
-from bravado.mapping.http_client import (Authenticator,
-                                         ApiKeyAuthenticator,
-                                         HttpClient)
-from bravado.requests_future import RequestsFuture
+from bravado.mapping.http_client import HttpClient
+from bravado.mapping.http_future import HttpFuture
+from bravado.mapping.response import ResponseLike
 
 log = logging.getLogger(__name__)
+
+
+class Authenticator(object):
+    """Authenticates requests.
+
+    :param host: Host to authenticate for.
+    """
+
+    def __init__(self, host):
+        self.host = host
+
+    def __repr__(self):
+        return u"%s(%s)" % (self.__class__.__name__, self.host)
+
+    def matches(self, url):
+        """Returns true if this authenticator applies to the given url.
+
+        :param url: URL to check.
+        :return: True if matches host, port and scheme, False otherwise.
+        """
+        split = urlparse.urlsplit(url)
+        return self.host == split.hostname
+
+    def apply(self, request):
+        """Apply authentication to a request.
+
+        :param request: Request to add authentication information to.
+        """
+        raise NotImplementedError(u"%s: Method not implemented",
+                                  self.__class__.__name__)
+
+
+# noinspection PyDocstring
+class ApiKeyAuthenticator(Authenticator):
+    """?api_key authenticator.
+
+    This authenticator adds a query parameter to specify an API key.
+
+    :param host: Host to authenticate for.
+    :param api_key: API key.
+    :param param_name: Query parameter specifying the API key.
+    """
+
+    def __init__(self, host, api_key, param_name=u'api_key'):
+        super(ApiKeyAuthenticator, self).__init__(host)
+        self.param_name = param_name
+        self.api_key = api_key
+
+    def apply(self, request):
+        request.params[self.param_name] = self.api_key
+        return request
 
 
 class BasicAuthenticator(Authenticator):
@@ -45,17 +94,22 @@ class RequestsClient(HttpClient):
         self.session = requests.Session()
         self.authenticator = None
 
-    def request(self, request_params, op=None):
+    def request(self, request_params, response_callback=None):
         """
         :param request_params: complete request data.
         :type request_params: dict
-        :return: request
-        :rtype: requests.Request
+        :param response_callback: Function to be called on the response
+        :returns: HTTP Future object
+        :rtype: :class: `bravado.mapping.http_future.HttpFuture`
         """
-        return RequestsFuture(
-            op,
-            self.session,
-            self.authenticated_request(request_params))
+        requests_future = RequestsFutureAdapter(
+            self.session, self.authenticated_request(request_params))
+
+        return HttpFuture(
+            requests_future,
+            RequestsResponseAdapter,
+            response_callback,
+            )
 
     def set_basic_auth(self, host, username, password):
         self.authenticator = BasicAuthenticator(
@@ -73,3 +127,72 @@ class RequestsClient(HttpClient):
             return self.authenticator.apply(request)
 
         return request
+
+
+def add_response_detail_to_errors(e):
+    """Specific to requests errors. Error detail is not
+    directly visible in `raise_for_status` trace, instead it is
+    located under `e.response.text`
+    """
+    if hasattr(e, 'response') and hasattr(e.response, 'text'):
+        # e.args is a tuple, change to list for modifications
+        args = list(e.args)
+        args[0] += (' : ' + e.response.text)
+        e.args = tuple(args)
+    raise e
+
+
+class RequestsResponseAdapter(ResponseLike):
+    """Wraps a requests.models.Response object to provide a uniform interface
+    to the response innards.
+    """
+
+    def __init__(self, requests_lib_response):
+        """
+        :type requests_lib_response: :class:`requests.models.Response`
+        """
+        self._delegate = requests_lib_response
+
+    @property
+    def status_code(self):
+        return self._delegate.status_code
+
+    def json(self, **kwargs):
+        return self._delegate.json(**kwargs)
+
+
+class RequestsFutureAdapter(object):
+    """A future which inputs HTTP params"""
+
+    def __init__(self, session, request):
+        """Kicks API call for Requests client
+
+        :param session: session object to use for making the request
+        :param request: dict containing API request parameters
+        """
+        self.session = session
+        self.request = request
+
+    def check_for_exceptions(self, response):
+        try:
+            response.raise_for_status()
+        except Exception as e:
+            add_response_detail_to_errors(e)
+
+    def result(self, timeout):
+        """Blocking call to wait for API response
+
+        :param timeout: timeout in seconds to wait for response
+        :type timeout: integer
+        :return: raw response from the server
+        :rtype: dict
+        """
+        request = self.request
+        log.debug(u"%s %s(%r)", request.method, request.url, request.params)
+        response = self.session.send(
+            self.session.prepare_request(request),
+            timeout=timeout)
+
+        self.check_for_exceptions(response)
+
+        return response
