+++ tests/mapping/operation/unmarshal_response_test.py
@@ -11,7 +11,7 @@ def test_no_content(empty_swagger_spec):
     response = Mock(spec=ResponseLike, status_code=200)
     with patch('bravado.mapping.operation.get_response_spec') as m:
         m.return_value = response_spec
-        op = Mock(swagger_spec=empty_swagger_spec)
+        op = Mock(spec=empty_swagger_spec)
         status_code, value = unmarshal_response(response, op)
     assert 200 == status_code
     assert value is None
@@ -24,7 +24,7 @@ def test_json_content(empty_swagger_spec):
             'type': 'string',
         }
     }
-    op = Mock(swagger_spec=empty_swagger_spec)
+    op = Mock(spec=empty_swagger_spec)
     response = Mock(
         spec=ResponseLike,
         status_code=200,
@@ -32,7 +32,7 @@ def test_json_content(empty_swagger_spec):
 
     with patch('bravado.mapping.operation.get_response_spec') as m:
         m.return_value = response_spec
-        op = Mock(swagger_spec=empty_swagger_spec)
+        op = Mock(spec=empty_swagger_spec)
         status_code, value = unmarshal_response(response, op)
     assert 200 == status_code
     assert 'Monday' == value
